gemini-test/requirements.txt
google-generativeai==0.5.0